requirements.txt
python-telegram-bot
requests
aiohttp
APScheduler
python-dotenv
pycoingecko